LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/navigate-same-document-expected.txt


FAIL navigate() navigates same-document and fires onnavigate (async) assert_equals: expected "#d" but got ""
FAIL navigate() navigates same-document and fires onnavigate (sync) assert_equals: expected "#d" but got ""
